Remove custom code for disabling AI integration
