+++ rest-narration/src/main/resources/application.properties
@@ -1,12 +1,9 @@
 quarkus.application.name=rest-narration
 quarkus.banner.path=banner.txt
 
-# Narration configuration
-%dev,test,openai,azure-openai.narration.make-live-calls=true
-narration.fallback-narration=High above a bustling city, a symbol of hope and justice soared through the sky, while chaos reigned below, with malevolent laughter echoing through the streets.\nWith unwavering determination, the figure swiftly descended, effortlessly evading explosive attacks, closing the gap, and delivering a decisive blow that silenced the wicked laughter.\n\nIn the end, the battle concluded with a clear victory for the forces of good, as their commitment to peace triumphed over the chaos and villainy that had threatened the city.\nThe people knew that their protector had once again ensured their safety.\n
-
 ## Default OpenAI configuration
 quarkus.langchain4j.chat-model.provider=openai
+quarkus.langchain4j.openai.enable-integration=false
 quarkus.langchain4j.openai.api-key=change-me
 quarkus.langchain4j.openai.embedding-model.enabled=false
 quarkus.langchain4j.openai.moderation-model.enabled=false
@@ -16,6 +13,7 @@ quarkus.langchain4j.openai.chat-model.temperature=0.7
 quarkus.langchain4j.openai.chat-model.top-p=0.5
 
 ## Default Azure OpenAI configuration
+quarkus.langchain4j.azure-openai.enable-integration=false
 quarkus.langchain4j.azure-openai.api-key=change-me
 quarkus.langchain4j.azure-openai.embedding-model.enabled=false
 quarkus.langchain4j.azure-openai.moderation-model.enabled=false
@@ -30,6 +28,7 @@ quarkus.langchain4j.azure-openai.chat-model.top-p=${quarkus.langchain4j.openai.c
 %dev,test,openai.quarkus.langchain4j.chat-model.provider=openai
 %dev,test.quarkus.langchain4j.openai.base-url=http://localhost:${quarkus.wiremock.devservices.port}/v1/
 %dev,test.quarkus.langchain4j.openai.timeout=3s
+%dev,test,openai,azure-openai.quarkus.langchain4j.openai.enable-integration=true
 %dev,test,openai,azure-openai.quarkus.langchain4j.openai.log-requests=true
 %dev,test,openai,azure-openai.quarkus.langchain4j.openai.log-responses=true
 %dev,test.quarkus.langchain4j.openai.max-retries=2
@@ -39,6 +38,7 @@ quarkus.langchain4j.azure-openai.chat-model.top-p=${quarkus.langchain4j.openai.c
 %azure-openai.quarkus.langchain4j.chat-model.provider=azure-openai
 %dev,test.quarkus.langchain4j.azure-openai.endpoint=http://localhost:${quarkus.wiremock.devservices.port}/v1/
 %dev,test,quarkus.langchain4j.azure-openai.timeout=3s
+%dev,test,openai,azure-openai.quarkus.langchain4j.azure-openai.enable-integration=true
 %dev,test,openai,azure-openai.quarkus.langchain4j.azure-openai.log-requests=true
 %dev,test,openai,azure-openai.quarkus.langchain4j.azure-openai.log-responses=true
 %dev,test.quarkus.langchain4j.azure-openai.max-retries=2
